Move transaction_operation_wrapper to operation_processor.
